% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clixo-function.R
\name{clixo}
\alias{clixo}
\title{Create Clique-Extracted Ontology (CliXO)}
\usage{
clixo(
  similarity,
  alpha = 0.01,
  beta = 0.5,
  feature_name = "feature",
  onto_prefix = "CliXO"
)
}
\arguments{
\item{similarity}{Feature similarity, a square matrix of numerics containing
feature-feature similarity measures.}

\item{alpha}{A numeric of a noise parameter. Please see
https://pubmed.ncbi.nlm.nih.gov/24932003/.}

\item{beta}{A numeric of a parameter which deals with missing edges. Please
see https://pubmed.ncbi.nlm.nih.gov/24932003/.}

\item{feature_name}{A character to annotate feature (source)-ontology
(target) relation in the resulting ontology.}

\item{onto_prefix}{A character that precedes the resulting ontology names.}
}
\value{
clique-extracted ontology, a data frame with rows for ontologies and
four columns for source, target, similarity, and relation. Feature (source)-
ontology (target) relation is annotated as 'feature' as defined by default
for \code{feature_name}, while ontology-ontology relation is annotated as
'is_a'. To differentiate between feature and ontology names, an
\code{onto_prefix} with ':' precedes an ontology name. All columns except
similarity are characters. Similarity (a numeric) is a minimum threshold by
which either features or ontologies (source) belong to an ontology (target).
}
\description{
This function create clique-extracted ontology from a similarity matrix.
}
\examples{

## Create input example
input=input_example()

## Run CliXO algorithm
ontology=clixo(input$similarity)
}
\keyword{clixo}
