% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_tsv_for_clixo-function.R
\name{create_tsv_for_clixo}
\alias{create_tsv_for_clixo}
\title{Create a TSV file for CliXO input}
\usage{
create_tsv_for_clixo(similarity, path, heading = FALSE)
}
\arguments{
\item{similarity}{A square matrix of numerics containing
feature-feature similarity measures.}

\item{path}{A character of target file path (do not include file
extension).}

\item{heading}{A logical indicating if heading is included or not. By
default, CliXO required no heading.}
}
\value{
output A tab-separated value (TSV) file containing similarity table
with rows for each pair of features and three columns for source, target,
and similarity.
}
\description{
This function create a tab-separated value (TSV) file from a similarity
matrix for CliXO input.
}
\examples{

## Create input example
input=input_example()

## Create a TSV file from a similarity matrix for CliXO input
create_tsv_for_clixo(input$similarity,'example')
}
\keyword{(TSV),}
\keyword{CliXO}
\keyword{input}
\keyword{tab-separated}
\keyword{value}
