% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_example-function.R
\name{input_example}
\alias{input_example}
\title{Make an input example for clixo package}
\usage{
input_example()
}
\value{
output A list of inputs: 1) value, a data frame with rows for
instances and columns for features; 2) similarity, a square matrix of
numerics containing feature-feature similarity measures.
}
\description{
This function create an input example for several function in clixo package.
}
\examples{

## Create input example
input=input_example()
}
\keyword{data}
\keyword{example}
